% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_alpha_mu}
\alias{getEpsilon_alpha_mu}
\alias{getEpsilon_alpha_mu,ZinbModel-method}
\title{Returns the regularization parameter for alpha_mu}
\usage{
getEpsilon_alpha_mu(object)

\S4method{getEpsilon_alpha_mu}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\description{
The regularization parameter is given by epsilon multiplied by
penalty_alpha_mu
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

