% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_beta_pi}
\alias{getEpsilon_beta_pi}
\alias{getEpsilon_beta_pi,ZinbModel-method}
\title{Returns the regularization parameter for beta_pi}
\usage{
getEpsilon_beta_pi(object)

\S4method{getEpsilon_beta_pi}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\description{
The regularization parameter is given by epsilon multiplied by
penalty_beta_pi
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

