% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_gamma_pi}
\alias{getEpsilon_gamma_pi}
\alias{getEpsilon_gamma_pi,ZinbModel-method}
\title{Returns the regularization parameter for gamma_pi}
\usage{
getEpsilon_gamma_pi(object)

\S4method{getEpsilon_gamma_pi}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\description{
The regularization parameter is given by epsilon multiplied by
penalty_gamma_pi
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

