% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_W}
\alias{getEpsilon_W}
\alias{getEpsilon_W,ZinbModel-method}
\title{Returns the regularization parameter for W}
\usage{
getEpsilon_W(object)

\S4method{getEpsilon_W}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\description{
The regularization parameter is given by epsilon multiplied by
penalty_W
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

