% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_zeta}
\alias{getEpsilon_zeta}
\alias{getEpsilon_zeta,ZinbModel-method}
\title{Returns the regularization parameter for the dispersion parameter}
\usage{
getEpsilon_zeta(object)

\S4method{getEpsilon_zeta}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\description{
The regularization parameter penalizes the variance of zeta, the log of 
the dispersion parameters across samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

