% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getJ}
\alias{getJ}
\title{Returns the number of genes}
\usage{
getJ(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving genes.}
}
\value{
the number of genes.
}
\description{
Given an object that describes a dataset or a model involving genes, this function returns the number of genes.
}
\examples{
a=zinb_model()
getJ(a) 
}

