% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getK}
\alias{getK}
\alias{getK,ZinbModel-method}
\title{Returns the number of latent factors}
\usage{
getK(object)

\S4method{getK}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving latent
factors}
}
\value{
the number of latent factors
}
\description{
Given an object that describes a dataset or a model involving latent factors,
this function returns the number of latent factors.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the number of latent factors.
}}
\examples{
a <- zinbModel()
getK(a) 
}

