% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getMu}
\alias{getMu}
\alias{getMu,ZinbModel-method}
\title{Returns the matrix of mean parameters}
\usage{
getMu(object)

\S4method{getMu}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the matrix of mean parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of mean parameters.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the mean of the non-zero component.
}}

\examples{
a <- zinbModel(n=5, J=10)
getMu(a) 
}
