% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-zinb_model.R
\docType{methods}
\name{getN}
\alias{getN}
\alias{getN,zinb_model-method}
\title{Returns the number of samples}
\usage{
getN(object)

\S4method{getN}{zinb_model}(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving samples.}
}
\value{
the number of samples
}
\description{
Given an object that describes a dataset or a model involving samples, this
function returns the number of samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{zinb_model}: return the number of samples.
}}
\examples{
a <- zinb_model()
getN(a) 
}

