% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getV_mu}
\alias{getV_mu}
\alias{getV_mu,ZinbModel-method}
\title{Returns the gene-level design matrix for mu}
\usage{
getV_mu(object, ...)

\S4method{getV_mu}{ZinbModel}(object, intercept = TRUE)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}

\item{intercept}{logical. Whether to return the intercept (ignored if V_mu
has no intercept).}
}
\value{
the gene-level design matrix for mu
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the gene-level design matrix for mu
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the gene-level design matrix for mu.
}}

\examples{
a <- zinbModel(n=5, J=10)
getV_mu(a) 
}
