% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getV_pi}
\alias{getV_pi}
\alias{getV_pi,ZinbModel-method}
\title{Returns the gene-level design matrix for pi}
\usage{
getV_pi(object, ...)

\S4method{getV_pi}{ZinbModel}(object, intercept = TRUE)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}

\item{intercept}{logical. Whether to return the intercept (ignored if V_pi has
no intercept).}
}
\value{
the gene-level design matrix for pi
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the gene-level design matrix for pi
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the sample-level design matrix for pi.
}}

