% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getW}
\alias{getW}
\alias{getW,ZinbModel-method}
\title{Returns the low-dimensional matrix of inferred sample-level covariates W}
\usage{
getW(object)

\S4method{getW}{ZinbModel}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{ZinbModel}} object, typically the result
of \code{\link{zinbFit}}.}
}
\value{
the matrix \code{W} of inferred sample-level covariates.
}
\description{
Given an object that contains the fit of a ZINB-WaVE model, returns the
matrix \code{W} of low-dimensional matrix of inferred sample-level
covariates.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

\examples{
a <- zinbModel(n=5, J=10)
getW(a) 
}
