% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getX_mu}
\alias{getX_mu}
\alias{getX_mu,ZinbModel-method}
\title{Returns the sample-level design matrix for mu}
\usage{
getX_mu(object, ...)

\S4method{getX_mu}{ZinbModel}(object, intercept = TRUE)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}

\item{intercept}{logical. Whether to return the intercept (ignored if X_mu
has no intercept). Default \code{TRUE}}
}
\value{
the sample-level design matrix for mu
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the sample-level design matrix for mu
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the sample-level design matrix for mu.
}}

