% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getX_pi}
\alias{getX_pi}
\alias{getX_pi,ZinbModel-method}
\title{Returns the sample-level design matrix for pi}
\usage{
getX_pi(object, ...)

\S4method{getX_pi}{ZinbModel}(object, intercept = TRUE)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}

\item{intercept}{logical. Whether to return the intercept (ignored if X_pi
has no intercept).}
}
\value{
the sample-level design matrix for pi
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the sample-level design matrix for pi
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the sample-level design matrix for pi.
}}

\examples{
a <- zinbModel(n=5, J=10)
getX_pi(a) 
}
