% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getZeta}
\alias{getZeta}
\alias{getZeta,ZinbModel-method}
\title{Returns the vector of log of inverse dispersion parameters}
\usage{
getZeta(object)

\S4method{getZeta}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the vector \code{zeta} of log of inverse dispersion parameters
}
\description{
Given an object that describes a matrix of zero-inflated negative binomial 
distributions, returns the vector \code{zeta} of log of inverse dispersion 
parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the log of the inverse of the dispersion parameter.
}}
\examples{
a <- zinbModel(n=5, J=10)
getZeta(a) 
}

