% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{grad}
\alias{grad}
\title{Gradient function of the zero-inflated negative binomial model}
\usage{
grad(parms, Y, Y0, Y1, X, W, kx, kw, offsetx, offsetz, linkobj)
}
\arguments{
\item{parms}{a vector of parameters: should contain the values of beta, followed by those of alpha, followed by the log(1/phi)}

\item{Y}{the data matrix (genes in rows, cells in columns)}

\item{Y0}{an indicator of Y==0}

\item{Y1}{an indicator of Y==1}

\item{X}{the design matrix for the regression on mu}

\item{W}{the design matrix for the regression on pi}

\item{kx}{the number of beta parameters}

\item{kw}{the number of alpha parameters}

\item{offsetx}{the offset for the regression on X}

\item{linkobj}{the link function object for the regression on pi (typically the result of binomial())}

\item{offsetw}{the offset for the regression on W}
}
\description{
This function computes the gradient of a standard regression model
}
\details{
This is a standard implementation of the gradient of a 
zero-inflated negative binomial regression model.
The problem with this implementation is that, when applied to our model, it requires
the design matrices to be (n*J) x p, where p = n * (k_W + 1) + J * (k_X + 1) + J,
n is the number of cells, J the number of genes, k_W is the number of covariates
of the logistic regression, k_X is the number of covariates in the log-linear regression.
}

