% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jp.R
\name{gradient.zinb.loglik.regression}
\alias{gradient.zinb.loglik.regression}
\title{Gradient of the log-likelihood of the zero-inflated negative binomial model for a regression model}
\usage{
gradient.zinb.loglik.regression(alpha, Y, X.mu = NULL, X.pi = NULL,
  X.theta = NULL, Y.mu = NULL, Y.pi = NULL, offset.mu = NULL,
  offset.pi = NULL, offset.theta = NULL, epsilon = 0)
}
\arguments{
\item{alpha}{the vectors of parameters c(a.mu, b, a.pi, a.theta) concatenated}

\item{Y}{the vector of counts}

\item{X.mu}{matrix of the model (see above, default=NULL)}

\item{X.pi}{matrix of the model (see above, default=NULL)}

\item{X.theta}{matrix of the model (see above, default=NULL)}

\item{Y.mu}{matrix of the model (see above, default=NULL)}

\item{Y.pi}{matrix of the model (see above, default=NULL)}

\item{offset.mu}{matrix of the model (see above, default=NULL)}

\item{offset.pi}{matrix of the model (see above, default=NULL)}

\item{offset.theta}{matrix of the model (see above, default=NULL)}
}
\description{
This function computes the gradient of the log-likelihood of a vector of counts according to a zero-inflated negative binomial model parametrized as follows:
Mean of the negative binomial: log(mu) = X.mu %*% a.mu + Y.mu %*%* b + offset.mu
Dispersion of the negative binomial: log(theta) = X.theta %*% a.theta + offset.theta
Probability of the zero component: logit(P(Y=0)) = X.pi %*% alpha.pi + Y.pi %*% b + offset.pi
Note that the b vector is shared between the mean of the negative binomial and the probability of zero.
}

