% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{kmeans_projection}
\alias{kmeans_projection}
\title{Function which runs k-means on the two-dimensional projection of the data  
Returns all outputs of kmeansruns function from fpc}
\usage{
kmeans_projection(data_kD, cl.range = NULL)
}
\arguments{
\item{data_kD}{matrix with cell's projections in k-dimensional space in lines, k is the number of columns of U}

\item{cl.range}{range for the possible number of clusters, default : from 2 to 10}
}
\description{
Function which runs k-means on the two-dimensional projection of the data  
Returns all outputs of kmeansruns function from fpc
}

