% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_smallmatrix.R
\name{loglikfull}
\alias{loglikfull}
\title{Log-likelihood function of a zero-inflated negative binomial model}
\usage{
loglikfull(parms, Y, X, W, kx, kw)
}
\arguments{
\item{parms}{a vector of parameters: should contain the values of beta, followed by those of alpha, followed by those of log(1/phi)}

\item{Y}{the data matrix (genes in rows, cells in columns)}

\item{X}{the design matrix for the regression on mu (n x k_X)}

\item{W}{the design matrix for the regression on pi (J x k_W)}

\item{kx}{the number of beta parameters}

\item{kw}{the number of alpha parameters}
}
\description{
This function computes the log-likelihood of a zero-inflated negative binomial model
}
\details{
This model let the user specify both a regression on mu and on pi. Note that pi will always depend on beta0
in addition to the covariates specified by the user. For this to be meaningful, we should consider a parametrization
of X for which beta0 is the overall mean (e.g., contrasts = "contr.sum").
}

