% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\docType{methods}
\name{nFactors,ZinbModel-method}
\alias{nFactors,ZinbModel-method}
\title{Return the number of latent factors}
\usage{
\S4method{nFactors}{ZinbModel}(x)
}
\arguments{
\item{x}{an object of class \code{ZinbModel}.}
}
\value{
the number of latent factors.
}
\description{
Given an object that describes a ZINB model involving latent factors,
this function returns the number of latent factors.
}
\examples{
a <- zinbModel()
nFactors(a)
}

