% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\docType{methods}
\name{nFeatures,ZinbModel-method}
\alias{nFeatures,ZinbModel-method}
\title{Return the number of features}
\usage{
\S4method{nFeatures}{ZinbModel}(x)
}
\arguments{
\item{x}{an object of class \code{ZinbModel}.}
}
\value{
the number of genes.
}
\description{
Given an object that describes a ZINB model, this
function returns the number of genes.
}
\examples{
a <- zinbModel()
nFeatures(a)
}

