% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\docType{methods}
\name{nSamples,ZinbModel-method}
\alias{nSamples,ZinbModel-method}
\title{Return the number of samples}
\usage{
\S4method{nSamples}{ZinbModel}(x)
}
\arguments{
\item{x}{an object of class \code{ZinbModel}.}
}
\value{
the number of samples
}
\description{
Given an object that describes a ZINB model, this function returns the number
of samples.
}
\examples{
a <- zinbModel()
nSamples(a) 
}

