% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jp.R
\name{orthogonalizeJointly}
\alias{orthogonalizeJointly}
\title{Orthogonalize U, V, W}
\usage{
orthogonalizeJointly(U, V, W)
}
\arguments{
\item{U}{left matrix}

\item{V}{first right matrix}

\item{W}{second right matrix}
}
\description{
Given U, V, W, find U2, V2, W2 such that UV'=U2V2' and
UW'=U2W2' and such that ||U2||^2+||V2||^2+||W||^2 is minimal.
}

