% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{penalty}
\alias{penalty}
\alias{penalty,zinb_model-method}
\title{Compute the penalty of a model}
\usage{
penalty(model)

\S4method{penalty}{zinb_model}(model)
}
\arguments{
\item{model}{an object that describes a statistical model with regularization
parameters.}
}
\value{
The penalty of the model.
}
\description{
Given a statistical model with regularization parameters, compute the
penalty.
}
\section{Methods (by class)}{
\itemize{
\item \code{zinb_model}: return the penalization.
}}
\examples{
m <- zinb_model(K=2)
penalty(m)
}

