% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{rand.measure}
\alias{rand.measure}
\title{Function which calculate rand indices comparing a reference partition to the zinb partition and  
to other partitions which are provided}
\usage{
rand.measure(ref = NULL, zinb.partition = NULL,
  matrix.of.partitions = NULL)
}
\arguments{
\item{ref}{reference partition}

\item{zinb.partition}{partition resulting from zinb}

\item{matrix}{which columns are partitions resulting from other methods}
}
\description{
Function which calculate rand indices comparing a reference partition to the zinb partition and  
to other partitions which are provided
}

