% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{silhouette.measure}
\alias{silhouette.measure}
\title{Function which calculate rand indices comparing a reference partition to the zinb partition and  
to other partitions which are provided}
\usage{
silhouette.measure(ref.clust, zinb.clust, add.clust)
}
\arguments{
\item{ref}{reference partition}

\item{zinb.partition}{partition resulting from zinb}

\item{matrix}{which columns are partitions resulting from other methods}
}
\description{
Function which calculate rand indices comparing a reference partition to the zinb partition and  
to other partitions which are provided
}

