% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sim.R
\name{simulateNB}
\alias{simulateNB}
\title{Simulate zero-inflated negative binomial matrices}
\usage{
simulateNB(X.mu = NULL, a.mu = NULL, U = NULL, V = NULL,
  offset.mu = NULL, X.pi = NULL, alpha.pi = NULL, W = NULL,
  offset.pi = NULL, X.theta = NULL, a.theta = NULL, offset.theta = NULL,
  sizefactors = NULL, verbose = NULL, no_cores = 1)
}
\arguments{
\item{X.mu}{matrix of the model (see above, default=NULL)}

\item{a.mu}{matrix of coefficients (see above, default=NULL)}

\item{U}{matrix of the model (see above, default=NULL)}

\item{V}{matrix of the model (see above, default=NULL)}

\item{offset.mu}{matrix of the model (see above, default=NULL)}

\item{X.pi}{matrix of the model (see above, default=NULL)}

\item{alpha.pi}{matrix of coefficients (see above, default=NULL)}

\item{W}{matrix of the model (see above, default=NULL)}

\item{offset.pi}{matrix of the model (see above, default=NULL)}

\item{X.theta}{matrix of the model (see above, default=NULL)}

\item{a.theta}{matrix of dispersion parameters (see above, default=NULL)}

\item{offset.theta}{matrix of the model (see above, default=NULL)}

\item{sizefactors}{is a vector of size factors of length equal to the number of cells}
}
\description{
Simulate negative binomial distribution according to the following model:
Mean of the negative binomial: log(mu) = X.mu %*% a.mu + U %*% V + offset.mu
Dispersion of the negative binomial: log(theta) = X.theta %*% a.theta + offset.theta
Probability of the zero component: logit(P(Y=0)) = X.pi %*% alpha.pi + U %*% W + offset.pi
}

