% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{simulateZINB}
\alias{simulateZINB}
\alias{simulateZINB,zinb_model-method}
\title{Simulate counts from a zero-inflated negative binomial model}
\usage{
simulateZINB(object, seed, ...)

\S4method{simulateZINB}{zinb_model}(object, seed, no_cores = 1)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated negative
binomial.}

\item{seed}{an optional integer to specify how the random number generator
should be initialized with a call to \code{set.seed}. If missing, the
random generator state is not changed.}

\item{...}{additional arguments.}
}
\value{
the matrix of probabilities of 0.
}
\description{
Given an object that describes zero-inflated negative binomial distribution,
simulate counts from the distribution.
}
\section{Methods (by class)}{
\itemize{
\item \code{zinb_model}: simulate from a ZINB distribution.
}}
\examples{
a <- zinb_model(n=5, J=10)
simulateZINB(a) 
}

