% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jp.R
\name{zinb.loglik}
\alias{zinb.loglik}
\title{Log-likelihood of the zero-inflated negative binomial model}
\usage{
zinb.loglik(Y, mu, theta, logitP0)
}
\arguments{
\item{Y}{the counts}

\item{mu}{the mean parameters of the negative binomial}

\item{theta}{the dispersion parameters of the negative binomial}

\item{logitP0}{the logit of the probability of the zero component}
}
\description{
Log-likelihood of the zero-inflated negative binomial model
}

