% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jp.R
\name{zinb.PCA.00}
\alias{zinb.PCA.00}
\title{Estimation of latent factors from a count matrix (slow version)}
\usage{
zinb.PCA.00(datamatrix, k = 2, alt.number = 25, epsilon = 0.1,
  stop.epsilon = 1e-04, verbose = FALSE, no_cores = 1)
}
\arguments{
\item{datamatrix}{the count data matrix (cells in rows, genes in columns)}

\item{k}{number of latent factors (default 2)}

\item{alt.number}{maximum number of iterations (default 25)}

\item{epsilon}{regularization parameter (default 0.1)}

\item{stop.epsilon}{stopping criterion, when the relative gain in likelihood is below epsilon (default 0.0001)}

\item{verbose}{print information (default FALSE)}

\item{no_cores}{number of cores (default to 1)}
}
\description{
Estimation of latent factors from a count matrix (slow version)
}

