% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_svetlana.R
\name{zinb}
\alias{zinb}
\title{Estimation of latent factors from a count matrix}
\usage{
zinb(datamatrix, alt.number = 25)
}
\arguments{
\item{datamatrix}{the count data matrix (cells in rows, genes in columns)}

\item{alt.number}{maximum number of iterations (default 25)}
}
\description{
Estimation of latent factors from a count matrix
}

