% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_svetlana.R
\name{zinb2}
\alias{zinb2}
\title{Estimation of latent factors from a count matrix}
\usage{
zinb2(datamatrix, k = 2, alt.number = 25, epsilon = 0.1, verbose = TRUE,
  use.old = FALSE)
}
\arguments{
\item{datamatrix}{the count data matrix (cells in rows, genes in columns)}

\item{k}{number of latent factors (default 2)}

\item{alt.number}{maximum number of iterations (default 25)}

\item{epsilon}{regularization parameter (default 0.1)}

\item{verbose}{print information (default TRUE)}
}
\description{
Estimation of latent factors from a count matrix
}

