% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-zinb_model.R
\name{zinb_model}
\alias{zinb_model}
\title{Initialize an object of class zinb_model}
\usage{
zinb_model(...)
}
\arguments{
\item{...}{arguments passed to \code{new()}. See the \code{slots} section in 
\code{\link{`zinb_model-class`}}.}
}
\description{
Initialize an object of class zinb_model
}
\details{
This is a light wrapper around the new() function to create an 
  instance of class \code{zinb_model}.

If any of the related matrices are passed, \code{n}, \code{J}, and
  \code{K} are inferred. Alternatively, the user can specify one or more of
  \code{n}, \code{J}, and \code{K}.

A call with no argument has the following default values: \code{n =
  50}, \code{J = 100}, \code{K = 2}.

Although it is possible to create new instances of the class by 
  calling this function, this is not the most common way of creating
  \code{zinb_model} objects. The main use of the class is within the
  \code{\link{zinbFit}} function.
}
\examples{
a <- zinb_model()
getN(a)
getJ(a)
getK(a)

}

