% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zinb.R
\name{zinb_optimize}
\alias{zinb_optimize}
\title{Optimize the parameters of a ZINB regression model}
\usage{
zinb_optimize(m, Y)
}
\arguments{
\item{m}{The model of class zinb_model}

\item{Y}{The matrix of counts.}
}
\value{
An object of class zinb_model similar to the one given as argument with modified parameters alpha_mu, alpha_pi, beta_mu, beta_pi, gamma_mu, gamma_pi, W.
}
\description{
The parameters of the model given as argument are optimized by penalized maximum likelihood on the count matrix given as argument. It is recommended to call zinb_initialize before this function to have good starting point for optimization, since the optimization problem is not convex and can only converge to a local minimum.
}
\examples{
Y = matrix(10,3,5)
m = zinb_model(n=NROW(Y),J=NCOL(Y))
m = zinb_initialize(m,Y)
m = zinb_optimize(m,Y)
}

