% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_fit.R
\name{zinbInitialize}
\alias{zinbInitialize}
\title{Initialize the parameters of a ZINB regression model}
\usage{
zinbInitialize(m, Y, nb.repeat = 2, ncores = 1)
}
\arguments{
\item{m}{The model of class ZinbModel}

\item{Y}{The matrix of counts.}

\item{nb.repeat}{Number of iterations for the estimation of beta_mu and
gamma_mu.}

\item{ncores}{The number of cores. To be passed to mclapply.}
}
\value{
An object of class ZinbModel similar to the one given as argument 
  with modified parameters alpha_mu, alpha_pi, beta_mu, beta_pi, gamma_mu, 
  gamma_pi, W.
}
\description{
The initialization performs quick optimization of the parameters with several
simplifying assumptions compared to the true model: non-zero counts are 
models as log-Gaussian, zeros are modeled as dropouts. The dispersion 
parameter is not modified.
}
\examples{
Y <- matrix(rpois(15, lambda=1), 3, 5)
m <- zinbModel(n=NCOL(Y), J=NROW(Y))
m <- zinbInitialize(m, Y)
}

