% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/helper_methods.R
\docType{class}
\name{ZinbModel-class}
\alias{ZinbModel}
\alias{ZinbModel-class}
\alias{nFactors,ZinbModel-method}
\alias{nFeatures,ZinbModel-method}
\alias{nSamples,ZinbModel-method}
\alias{show,ZinbModel-method}
\title{Class ZinbModel}
\usage{
\S4method{show}{ZinbModel}(object)

\S4method{nSamples}{ZinbModel}(x)

\S4method{nFeatures}{ZinbModel}(x)

\S4method{nFactors}{ZinbModel}(x)
}
\arguments{
\item{object}{an object of class \code{ZinbModel}.}

\item{x}{an object of class \code{ZinbModel}.}
}
\description{
Objects of this class store all the values needed to work with a
zero-inflated negative binomial (ZINB) model, as described in the vignette.
They contain all information to fit a model by penalized maximum likelihood
or simulate data from a model.
}
\details{
For the full description of the model see the model vignette.
  Internally, the slots are checked so that the matrices are of the
  appropriate dimensions: in particular, \code{X}, \code{O_mu}, \code{O_pi},
  and \code{W} need to have \code{n} rows, \code{V} needs to have \code{J}
  rows, \code{zeta} must be of length \code{J}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show useful info on the object.

\item \code{nSamples}: returns the number of samples.

\item \code{nFeatures}: returns the number of features.

\item \code{nFactors}: returns the number of latent factors.
}}
\section{Slots}{

\describe{
\item{\code{X}}{matrix. The design matrix containing sample-level covariates, one
sample per row.}

\item{\code{V}}{matrix. The design matrix containing gene-level covariates, one gene
per row.}

\item{\code{O_mu}}{matrix. The offset matrix for mu.}

\item{\code{O_pi}}{matrix. The offset matrix for pi.}

\item{\code{which_X_mu}}{integer. Indeces of which columns of X to use in the
regression of mu.}

\item{\code{which_V_mu}}{integer. Indeces of which columns of V to use in the
regression of mu.}

\item{\code{which_X_pi}}{integer. Indeces of which columns of X to use in the
regression of pi.}

\item{\code{which_V_pi}}{integer. Indeces of which columns of V to use in the
regression of pi.}

\item{\code{X_mu_intercept}}{logical. TRUE if X_mu contains an intercept.}

\item{\code{X_pi_intercept}}{logical. TRUE if X_pi contains an intercept.}

\item{\code{V_mu_intercept}}{logical. TRUE if V_mu contains an intercept.}

\item{\code{V_pi_intercept}}{logical. TRUE if V_pi contains an intercept.}

\item{\code{W}}{matrix. The factors of sample-level latent factors.}

\item{\code{beta_mu}}{matrix or NULL. The coefficients of X in the regression of mu.}

\item{\code{gamma_mu}}{matrix or NULL. The coefficients of V in the regression of mu.}

\item{\code{alpha_mu}}{matrix or NULL. The coefficients of W in the regression of mu.}

\item{\code{beta_pi}}{matrix or NULL. The coefficients of X in the regression of pi.}

\item{\code{gamma_pi}}{matrix or NULL. The coefficients of V in the regression of pi.}

\item{\code{alpha_pi}}{matrix or NULL. The coefficients of W in the regression of pi.}

\item{\code{zeta}}{numeric. A vector of log of inverse dispersion parameters.}

\item{\code{epsilon_beta_mu}}{nonnegative scalar. Regularization parameter for
beta_mu}

\item{\code{epsilon_gamma_mu}}{nonnegative scalar. Regularization parameter for
gamma_mu}

\item{\code{epsilon_beta_pi}}{nonnegative scalar. Regularization parameter for
beta_pi}

\item{\code{epsilon_gamma_pi}}{nonnegative scalar. Regularization parameter for
gamma_pi}

\item{\code{epsilon_W}}{nonnegative scalar. Regularization parameter for W}

\item{\code{epsilon_alpha}}{nonnegative scalar. Regularization parameter for alpha
(both alpha_mu and alpha_pi)}

\item{\code{epsilon_zeta}}{nonnegative scalar. Regularization parameter for zeta}

\item{\code{epsilon_min_logit}}{scalar. Minimum regularization parameter for
parameters of the logit model, including the intercept.}
}}

