% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_fit.R
\name{zinbOptimizeDispersion}
\alias{zinbOptimizeDispersion}
\title{Optimize the dispersion parameters of a ZINB regression model}
\usage{
zinbOptimizeDispersion(m, Y, commondispersion = TRUE, ncores = 1)
}
\arguments{
\item{m}{The model of class ZinbModel}

\item{Y}{The matrix of counts.}

\item{commondispersion}{Whether or not a single dispersion for all features
is estimated (default TRUE)}

\item{ncores}{Number of cores for parallel computation (default 1)}
}
\value{
An object of class ZinbModel similar to the one given as argument 
  with modified parameters zeta.
}
\description{
The dispersion parameters of the model given as argument are optimized by 
penalized maximum likelihood on the count matrix given as argument.
}
\examples{
Y = matrix(10,3,5)
m = zinbModel(n=NROW(Y),J=NCOL(Y))
m = zinbInitialize(m,Y)
m = zinbOptimizeDispersion(m,Y)
}

