% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_svetlana.R
\name{ziNegBin}
\alias{ziNegBin}
\title{Log-likelihood function of the zero-inflated negative binomial model for the estimation of "right parts" with 
log(M)=X.M*alpha.M+U*V; logit(Pi)=X.pi*alpha.pi+U*W; M=E[Y] and Y data matrix}
\usage{
ziNegBin(parms, j, X.M = NULL, X.pi = NULL, U, Y, epsilon = 0,
  offsetx = 0, offsetz = 0)
}
\arguments{
\item{parms}{a vector of parameters: j-th column of alpha.M, followed by j-th column of V, followed by j-th column of alpha.pi, followed by j-th column of W, followed by the log(1/phi)}

\item{j}{is the number of column (gene) to be optimized over}

\item{X.M}{is the design matrix for M}

\item{X.pi}{is the design matrix for Pi.}

\item{U}{is the current value for the latent factor matrix}

\item{Y}{the data matrix (cells in rows, genes in columns)}

\item{offsetx}{the offset for the regression on M}

\item{offsetw}{the offset for the regression on Pi}

\item{linkobj}{the link function object for the regression on pi (typically the result of binomial())}
}
\description{
Log-likelihood function of the zero-inflated negative binomial model for the estimation of "right parts" with 
log(M)=X.M*alpha.M+U*V; logit(Pi)=X.pi*alpha.pi+U*W; M=E[Y] and Y data matrix
}

