% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_svetlana.R
\name{ziNegBin.U}
\alias{ziNegBin.U}
\title{Log-likelihood function of the zero-inflated negative binomial model for the estimation of U in the model 
log(M)=X.M*alpha.M+U*V; logit(Pi)=X.pi*alpha.pi+U*W; M=E[Y] with Y data matrix}
\usage{
ziNegBin.U(parms, i, V, W, X.M = F, X.pi = F, alpha.M = F, alpha.pi = F,
  theta, Y)
}
\arguments{
\item{parms}{a vector of parameters: i-th line of U}

\item{i}{is the number of ligne (cell) to be optimized over}

\item{V, W, alpha.M, alpha.pi}{are the "right parts" to be given as arguments}

\item{X.M}{is the design matrix for M}

\item{X.pi}{is the design matrix for Pi.}

\item{theta}{vector of length J with gene specific dispersion parameters}

\item{Y}{the data matrix (cells in rows, genes in columns)}

\item{linkobj}{the link function object for the regression on pi (typically the result of binomial())}
}
\description{
Log-likelihood function of the zero-inflated negative binomial model for the estimation of U in the model 
log(M)=X.M*alpha.M+U*V; logit(Pi)=X.pi*alpha.pi+U*W; M=E[Y] with Y data matrix
}

