% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getEpsilon_beta_pi}
\alias{getEpsilon_beta_pi}
\alias{getEpsilon_beta_pi,ZinbModel-method}
\title{Returns the vector of regularization parameter for beta_pi}
\usage{
getEpsilon_beta_pi(object)

\S4method{getEpsilon_beta_pi}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{beta_pi}.
}
\description{
Given an object describing a ZINB model, returns a vector of size the number
of rows in the parameter \code{beta_pi} with the regularization parameters
associated to each row.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: method for ZinbModel.
}}

\examples{
a <- zinbModel(n=5, J=10)
getEpsilon_beta_pi(a) 
}
