% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-zinb_model.R
\docType{methods}
\name{getJ}
\alias{getJ}
\alias{getJ,zinb_model-method}
\title{Returns the number of genes}
\usage{
getJ(object)

\S4method{getJ}{zinb_model}(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving genes.}
}
\value{
the number of genes.
}
\description{
Given an object that describes a dataset or a model involving genes, this
function returns the number of genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{zinb_model}: return the number of genes.
}}
\examples{
a <- zinb_model()
getJ(a) 
}

