% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getK}
\alias{getK}
\title{Returns the number of latent factors}
\usage{
getK(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving latent
factors}
}
\value{
the number of latent factors
}
\description{
Given an object that describes a dataset or a model involving latent factors,
this function returns the number of latent factors.
}
\examples{
a <- zinb_model()
getK(a) 
}

