% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getLogitPi}
\alias{getLogitPi}
\alias{getLogitPi,ZinbModel-method}
\title{Returns the matrix of logit of probabilities of zero}
\usage{
getLogitPi(object)

\S4method{getLogitPi}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the matrix of logit-probabilities of 0
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of logit of probabilities of 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the logit-probability of zero.
}}

\examples{
a <- zinbModel(n=5, J=10)
getLogitPi(a) 
}
