% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getMu}
\alias{getMu}
\title{Returns the matrix of mean parameters}
\usage{
getMu(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated distributions.}
}
\value{
the matrix of mean parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions, returns the matrix of mean parameters
}
\examples{
a=zinb_model(n=5,J=10)
getMu(a) 
}

