% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getN}
\alias{getN}
\title{Returns the number of samples}
\usage{
getN(object)
}
\arguments{
\item{object}{an object that describes a dataset or a model involving samples.}
}
\value{
the number of samples
}
\description{
Given an object that describes a dataset or a model involving samples, this function returns the number of samples
}
\examples{
a=zinb_model()
getN(a) 
}

