% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getPhi}
\alias{getPhi}
\alias{getPhi,ZinbModel-method}
\title{Returns the vector of dispersion parameters}
\usage{
getPhi(object)

\S4method{getPhi}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated.
distributions.}
}
\value{
the vector of dispersion parameters
}
\description{
Given an object that describes a matrix of zero-inflated negative binomial
distributions, returns the vector of dispersion parameters \code{phi}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the dispersion parameter.
}}

\examples{
a <- zinbModel(n=5, J=10)
getPhi(a) 
}
