% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getPi}
\alias{getPi}
\title{Returns the matrix of probabilities of zero}
\usage{
getPi(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the matrix of probabilities of 0
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of probabilities of 0.
}
\examples{
a <- zinb_model(n=5, J=10)
getPi(a) 
}

