% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{getTheta}
\alias{getTheta}
\alias{getTheta,ZinbModel-method}
\title{Returns the vector of inverse dispersion parameters}
\usage{
getTheta(object)

\S4method{getTheta}{ZinbModel}(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the vector of inverse dispersion parameters theta
}
\description{
Given an object that describes a matrix of zero-inflated negative binomial 
distributions, returns the vector of inverse dispersion parameters
\code{theta}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ZinbModel}: return the inverse of the dispersion parameter.
}}
\examples{
a <- zinbModel(n=5, J=10)
getTheta(a) 
}

