% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loglik}
\alias{loglik}
\title{Compute the log-likelihood of a model given some data}
\usage{
loglik(model, x, ...)
}
\arguments{
\item{model}{an object that describes a statistical model.}

\item{x}{an object that describes data.}
}
\value{
The log-likelihood of the model given the data.
}
\description{
Given a statistical model and some data, this function computes the
log-likelihood of the model given the data, i.e., the log-probability of the
data under the model.
}
\examples{
m <- zinb_model(n=5, J=10)
x <- simulateZINB(m)
loglik(m, x$counts)
}

