% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{penalty}
\alias{penalty}
\title{Compute the penalty of a model}
\usage{
penalty(model)
}
\arguments{
\item{model}{an object that describes a statistical model with regularization parameters}
}
\value{
The penalty of the model.
}
\description{
Given a statistical model with regularization parameters, compute the penalty.
}
\examples{
m=zinb_model(K=2)
penalty(m)
}

