% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simulateZINB}
\alias{simulateZINB}
\title{Simulate counts from a zero-inflated negative binomial model}
\usage{
simulateZINB(object, seed, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated negative binomial}

\item{seed}{an optional integer to specify how the random number generator should be initialized with a call to \code{set.seed}. If missing, the random generator state is not changed.}
}
\value{
the matrix of probabilities of 0
}
\description{
Given an object that describes zero-inflated negative binomial distribution, simulate counts from the distribution.
}
\examples{
a=zinb_model(n=5,J=10)
simulateZINB(a) 
}

