% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zinb_fit.R
\name{zinbFit}
\alias{zinbFit}
\title{Fit a ZINB regression model}
\usage{
zinbFit(Y, ...)

zinbFit(Y, ...)
}
\arguments{
\item{Y}{The data.}

\item{...}{Additional parameters to describe the model.}

\item{Y}{The matrix of counts.}

\item{...}{Additional parameters to describe the model}
}
\value{
An object of class \code{ZinbModel} that has been fitted by penalized
  maximum likelihood on the data.

An object of class ZinbModel that has been fitted by penalized maximum likelihood on the count matrix.
}
\description{
Given an object with the data, it fits a ZINB model.

Fit a ZINB regression model
}
\examples{
m=zinbFit(matrix(10,3,5))
}

