% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zinb_fit.R
\docType{methods}
\name{zinbFit}
\alias{zinbFit}
\alias{zinbFit,matrix-method}
\title{Fit a ZINB regression model}
\usage{
zinbFit(Y, ...)

\S4method{zinbFit}{matrix}(Y, ncores = 1, ...)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns).}

\item{...}{Additional parameters to describe the model, see \link{zinbModel}.}
}
\value{
An object of class \code{ZinbModel} that has been fitted by penalized
  maximum likelihood on the data.
}
\description{
Given an object with the data, it fits a ZINB model.
}
\details{
By default, i.e., if no arguments other than \code{Y} are passed, 
  the model is fitted with an intercept for the regression across-samples and
  one intercept for the regression across genes, both for mu and for pi.

This means that by default the model is fitted with \code{X_mu = 
  X_pi = 1_n} and \code{V_mu = V_pi = 1_J}. If the user explicitly passes the
  design matrices, this behavior is overwritten, i.e., the user needs to 
  explicitly include the intercept in the design matrices.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Y is a matrix of counts (genes in rows).
}}
\examples{
bio <- gl(2, 3)
m <- zinbFit(matrix(10, 10, 6), X=model.matrix(~bio))
}
\seealso{
\code{\link[stats]{model.matrix}}.
}

