% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_zinb.R
\name{zinbFit}
\alias{zinbFit}
\title{Fit a ZINB regression model}
\usage{
zinbFit(Y, ...)
}
\arguments{
\item{Y}{The matrix of counts.}

\item{...}{Additional parameters to describe the model}
}
\value{
An object of class ZinbModel that has been fitted by penalized maximum likelihood on the count matrix.
}
\description{
Fit a ZINB regression model
}
\examples{
m=zinbFit(matrix(10,3,5))
}

