% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{Class}
\name{ZinbModel-class}
\alias{ZinbModel-class}
\title{Class ZinbModel}
\description{
Objects of this class store all the values needed to work with a fitted 
model. Starting with an object of this class, one should be able to simulate 
data from a zero-inflated negative binomial and to compute the 
log-likelihood.
}
\details{
For the full description of the model see the model vignette. 
  Internally, the slots are checked so that the matrices are of the 
  appropriate dimensions: in particular, \code{X}, \code{O_mu}, \code{O_pi},
  and \code{W} need to have \code{n} rows, \code{V} needs to have \code{J}
  rows, \code{logtheta} must be of length \code{J}.
}
\section{Slots}{

\describe{
\item{\code{X}}{matrix. The design matrix containing sample-level covariates.}

\item{\code{V}}{matrix. The design matrix containing gene-level covariates.}

\item{\code{O_mu}}{matrix. The offset matrix for mu.}

\item{\code{O_pi}}{matrix. The offset matrix for pi.}

\item{\code{which_X_mu}}{integer. Indeces of which columns of X to use in the 
regression of mu.}

\item{\code{which_V_mu}}{integer. Indeces of which columns of V to use in the 
regression of mu.}

\item{\code{which_X_pi}}{integer. Indeces of which columns of X to use in the 
regression of pi.}

\item{\code{which_V_pi}}{integer. Indeces of which columns of V to use in the 
regression of pi.}

\item{\code{X_mu_intercept}}{logical. TRUE if X_mu contains an intercept.}

\item{\code{X_pi_intercept}}{logical. TRUE if X_pi contains an intercept.}

\item{\code{V_mu_intercept}}{logical. TRUE if V_mu contains an intercept.}

\item{\code{V_pi_intercept}}{logical. TRUE if V_pi contains an intercept.}

\item{\code{W}}{matrix. The factors of gene-level latent factors.}

\item{\code{beta_mu}}{matrix or NULL. The coefficients of X in the regression of mu.}

\item{\code{gamma_mu}}{matrix or NULL. The coefficients of V in the regression of mu.}

\item{\code{alpha_mu}}{matrix or NULL. The coefficients of W in the regression of mu.}

\item{\code{beta_pi}}{matrix or NULL. The coefficients of X in the regression of pi.}

\item{\code{gamma_pi}}{matrix or NULL. The coefficients of V in the regression of pi.}

\item{\code{alpha_pi}}{matrix or NULL. The coefficients of W in the regression of pi.}

\item{\code{logtheta}}{numeric. A vector of log of inverse dispersion parameters.}

\item{\code{epsilon}}{numeric. The regularization parameter for penalized maximum
likelihood estimation}

\item{\code{penalty_alpha_mu}}{numeric. This is a scalar that multiplies
epsilon to allow differential regularization for the rows of alpha_mu
parameters.}

\item{\code{penalty_beta_mu}}{numeric. This is a vector that multiplies
epsilon to allow differential regularization for the rows of beta_mu
parameters.}

\item{\code{penalty_gamma_mu}}{numeric. This is a vector that multiplies
epsilon to allow differential regularization for the rows of gamma_mu
parameters.}

\item{\code{penalty_alpha_pi}}{numeric. This is a scalar that multiplies
epsilon to allow differential regularization for the matrix of alpha_pi
parameters.}

\item{\code{penalty_beta_pi}}{numeric. This is a vector that multiplies
epsilon to allow differential regularization for the rows of beta_pi
parameters.}

\item{\code{penalty_gamma_pi}}{numeric. This is a vector that multiplies
epsilon to allow differential regularization for the rows of gamma_pi
parameters.}

\item{\code{penalty_W}}{numeric. This is a vector that multiplies epsilon to
allow differential regularization for the columns of W parameters.}

\item{\code{epsilon_varphi}}{numeric. This is the penalty for the variance of the
dispersion parameter.}
}}

